\name{QTLNetworkR-package}
\alias{QTLNetworkR-package}
\alias{QTLNetworkR}
\docType{package}
\title{
  Interactive software package for QTL visualization
}
\description{
  QTLNetworkR is a R package for Visualizing QTL mapping results generated by QTLNetwork. 
}
\details{
\tabular{ll}{
Package: \tab QTLNetworkR\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2009-10-07\cr
Depends: \tab utils, lattice, grid, gWidgets, gWidgetsRGtk2\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
URL: \tab \url{http://rqtlnetwork.r-forge.r-project.org/} \cr
}
Users can use library(QTLNetworkR) to start the GUI. Please Refer to the documentation
section of the project homepage for the usage of RQTLNetwork.
}
\author{
Zheng,Wenjun
Maintainer: Zheng,Wenjun <wjzh@zju.edu.cn>
}
%%\references{
%%QTLNetwork
%%}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{
R
QTLNetwork
QTL mapping
package
}
\examples{
library(QTLNetworkR)
}
